PlatformShims DRYed
