+++ utest/jvm/src-0/utest/PlatformShims.scala
@@ -6,16 +6,9 @@ import concurrent.duration._
 /**
  * Platform specific stuff that differs between JVM and JS
  */
-object PlatformShims {
-  def await[T](f: Future[T]): T = Await.result(f, 10.hours)
-
+trait PlatformShimsVersionSpecific {
   type EnableReflectiveInstantiation =
     utest.EnableReflectiveInstantiation
 
-  def loadModule(name: String, loader: ClassLoader): Any =
-    PortableScalaReflectExcerpts
-      .lookupLoadableModuleClass(name + "$", loader)
-      .getOrElse(throw new ClassNotFoundException(name))
-      .loadModule()
-
+  val Reflect = PortableScalaReflectExcerpts
 }
